refactoring state management/adding Context API
